Call parseUnsafeRelativePath() on non-relative paths.
While attempting to use httpx I found that it was not able to process
URLs like "https://127.0.0.1/%25", this is because the base golang url
parser will discard those chars as invalid.

Fortunately, it seems as if this library has another mechanism that can
be called for these specific cases, but unfortunately this function was
not being called for non-relative URLs.

This attempts to fix that issue but calling parseUnsafeRelativePath when
non-relative and only when the unsafe boolean is passed to ParseURL.